Rename folder to IPyWidgets
