+++ src/kernels/ipywidgets/baseIPyWidgetScriptManager.ts
