+++ src/kernels/ipywidgets/cdnWidgetScriptSourceProvider.ts
