+++ src/kernels/ipywidgets/ipyWidgetScriptManagerFactory.node.ts
