+++ src/kernels/ipywidgets/ipyWidgetScriptManagerFactory.web.ts
