+++ src/kernels/ipywidgets/ipyWidgetScriptSource.ts
