+++ src/kernels/ipywidgets/ipyWidgetScriptSourceProvider.ts
