+++ src/kernels/ipywidgets/localIPyWidgetScriptManager.node.ts
