+++ src/kernels/ipywidgets/localWidgetScriptSourceProvider.node.ts
