+++ src/kernels/ipywidgets/nbExtensionsPathProvider.node.ts
