+++ src/kernels/ipywidgets/nbExtensionsPathProvider.web.ts
