+++ src/kernels/ipywidgets/remoteIPyWidgetScriptManager.ts
