+++ src/kernels/ipywidgets/remoteWidgetScriptSourceProvider.ts
