+++ src/kernels/ipywidgets/scriptSourceProviderFactory.node.ts
