+++ src/kernels/ipywidgets/scriptSourceProviderFactory.web.ts
