+++ src/kernels/ipywidgets/scriptUriConverter.ts
