+++ src/kernels/ipywidgets/serviceRegistry.web.ts
