+++ src/webviews/webview-side/ipywidgets/common/manager.ts
@@ -16,7 +16,7 @@ import { logMessage, setLogger } from '../../react-common/logger';
 import { IMessageHandler, PostOffice } from '../../react-common/postOffice';
 import { create as createKernel } from './kernel';
 import { IIPyWidgetManager, IJupyterLabWidgetManager, IJupyterLabWidgetManagerCtor, ScriptLoader } from './types';
-import { WIDGET_MIMETYPE } from '../../../../kernels/ipywidgets-message-coordination/constants';
+import { WIDGET_MIMETYPE } from '../../../../kernels/ipywidgets/constants';
 import { KernelSocketOptions } from '../../../../kernels/types';
 import { Deferred, createDeferred } from '../../../../platform/common/utils/async';
 import {
