+++ src/webviews/webview-side/ipywidgets/common/requirejsRegistry.ts
@@ -3,7 +3,7 @@
 
 'use strict';
 
-import { WidgetScriptSource } from '../../../../kernels/ipywidgets-message-coordination/types';
+import { WidgetScriptSource } from '../../../../kernels/ipywidgets/types';
 import { logMessage } from '../../react-common/logger';
 
 type NonPartial<T> = {
